can save links into users folders and save comments comments save userID but not LinkID yet built search func to query flickr by a photos ID needs more testing

+++ server/api/link/link.model.js
@@ -15,7 +15,7 @@ module.exports = function(sequelize, DataTypes) {
         notEmpty: true
       }
     },//Unique?
-    apiID: DataTypes.INTEGER(11),
+    apiID: DataTypes.BIGINT,
     apiName: DataTypes.CHAR(45),
     publicComments: DataTypes.STRING(8000),
     active: DataTypes.BOOLEAN()
